+++ book/pages/ecosystem.md
@@ -1,2 +1,7 @@
+# More examples, tutorials
+
+This page, contains a list of additional axum-related crates, project showcases, and several tutorials taken from the `ECOSYSTEM.md` file.
+
+
 {{#include ../../ECOSYSTEM.md }}
