+++ book/pages/goals.md
@@ -4,7 +4,7 @@
 * A way to learn using axum.
 * Improve the examples (adding test, filling holes if there are any).
 * Adding `README.md` files to the examples.
-* The book can be found on https://axum.code-maven.com/ with the hope that one day it will be accepted as part of the official axum project
+* The book can be found at [axum.code-maven.com](https://axum.code-maven.com/) with the hope that one day it will be accepted as part of the official axum project
 and displayed on a stand-alone axum website or on the [Tokio](https://tokio.rs/) website.
