+++ book/pages/introduction.md
@@ -2,5 +2,5 @@
 
 The first steps writing a web application using axum.
 
-Simple handling of parameters in GET and POST request and as part of the path.
+Simple handling of parameters in GET and POST requests and as part of the path.
