--- book/pages/more-examples.md
@@ -1,4 +1,0 @@
-# More examples, tutorials
-
-The [ECOSYSTEM](./ecosystem.md) page contains a list of additional axum-related crates,
-project showcases, and several tutorials.
