+++ examples/hello-world/src/main.rs
@@ -28,24 +28,5 @@ async fn handler() -> Html<&'static str> {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use axum::{body::Body, http::Request, http::StatusCode};
-    use http_body_util::BodyExt;
-    use tower::ServiceExt;
+mod tests;
 
-    #[tokio::test]
-    async fn test_main_page() {
-        let response = app()
-            .oneshot(Request::builder().uri("/").body(Body::empty()).unwrap())
-            .await
-            .unwrap();
-
-        assert_eq!(response.status(), StatusCode::OK);
-        let body = response.into_body();
-        let bytes = body.collect().await.unwrap().to_bytes();
-        let html = String::from_utf8(bytes.to_vec()).unwrap();
-
-        assert_eq!(html, "<h1>Hello, World!</h1>");
-    }
-}
